requirements_beluga.txt
# local package
-e .

# external requirements
click==8.1.7
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

timm==0.9.10
tokenizers==0.19.1
torch>=2.2.0
torchmetrics==1.6.0
torchvision>=0.16.0
transformers==4.40.1
pytorch_lightning==2.4.0

decord==0.6.0
einops==0.6.1
h5py==3.12.0
huggingface_hub==0.23.4
hydra_core==1.3.2
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
omegaconf==2.3.0
openai==1.51.2
opencv-python==4.6.0.66
pandas==2.2.2
pillow==10.4.0
scikit_learn==1.5.0
scipy==1.14.1
sentencepiece==0.1.99
hydra_core==1.3.2
